update mbed-os and azure iot device sdk.
- mbed-os 7be79f93feff724cbd7ee88c7afdecc0b01559db, Jan 15, 2018
- azure iot device sdk 3035fd9adb045312b465b911bf58495f6ed6cf9c, Jan 11, 2018

+++ azure_c_shared_utility/uws_client.c
@@ -846,7 +846,7 @@ static int ParseStringToDecimal(const char *src, int* dst)
     int result;
     char* next;
 
-    (*dst) = strtol(src, &next, 0);
+    (*dst) = (int)strtol(src, &next, 0);
     if ((src == next) || ((((*dst) == INT_MAX) || ((*dst) == INT_MIN)) && (errno != 0)))
     {
         result = __FAILURE__;
@@ -1785,7 +1785,7 @@ int uws_client_send_frame_async(UWS_CLIENT_HANDLE uws_client, unsigned char fram
                             free(ws_pending_send);
                         }
 
-			            result = __FAILURE__;
+                        result = __FAILURE__;
                     }
                     else
                     {
